python/code_challenges/hashtable/challenge04/challenge04.py
class Node:
    def __init__(self, data):
        """
        Initializes a Node with the given data and sets the next node to None.
        """
        self.data = data
        self.next = None

class LinkedList:
    def __init__(self):
        """
        Initializes an empty LinkedList with the head set to None.
        """
        self.head = None

    def add(self, data):
        """
        Adds a node with the given data to the end of the LinkedList.
        """
        node = Node(data)
        if not self.head:
            self.head = node
        else:
            current = self.head
            while current.next:
                current = current.next
            current.next = node

    def find(self, key):
        """
        Finds and returns the node with the given key in the LinkedList. 
        Returns None if the key is not found.
        """
        current = self.head
        while current:
            if current.data[0] == key:
                return current
            current = current.next
        return None

    def __iter__(self):
        """
        Allows iteration over the LinkedList, yielding the data of each node.
        """
        current = self.head
        while current:
            yield current.data
            current = current.next

    def __str__(self):
        """
        Returns a string representation of the LinkedList.
        """
        return str(list(self))

class HashTable:
    def __init__(self, size=10):
        """
        Initializes a HashTable with the given size. Default size is 10.
        """
        self.size = size
        self.map = [None] * size

    def hash_key(self, key):
        """
        Generates a hash for the given key using the Python standard hash function.
        """
        return hash(key) % self.size

    def set(self, key, value):
        """
        Sets the key-value pair in the HashTable. Handles collisions using chaining.
        """
        hashed_key = self.hash_key(key)
        if self.map[hashed_key] is None:
            self.map[hashed_key] = LinkedList()
        # Add the data to the linked list at the hashed key
        existing_node = self.map[hashed_key].find(key)
        if existing_node:
            existing_node.data[1].append(value)
        else:
            self.map[hashed_key].add([key, [value]])

    def get(self, key):
        """
        Retrieves the value for the given key from the HashTable.
        """
        hashed_key = self.hash_key(key)
        bucket = self.map[hashed_key]
        if bucket is None:
            return None
        else:
            node = bucket.find(key)
            if node:
                return node.data[1]
            return None

def reorder_people_by_height(names, heights):
    # Create a hash table and add names and their corresponding heights
    hash_table = HashTable(len(names))
    for name, height in zip(names, heights):
        hash_table.set(height, name)

    # Sort the heights in descending order
    sorted_heights = sorted(set(heights), reverse=True)

    # Retrieve names from the hash table in the order of sorted heights
    sorted_names = []
    for height in sorted_heights:
        names_at_height = hash_table.get(height)
        if names_at_height:
            sorted_names.extend(names_at_height)

    return sorted_names

if __name__ == "__main__":
    names = ["Person1", "Person2", "Person3", "Person4", "Person5"]
    heights = [160, 180, 150, 170, 175]
    print(reorder_people_by_height(names, heights))